ajustes y formulario agregar productos

+++ app/Views/productos/index.php
@@ -1,11 +1,12 @@
 <?php
+helper('number');
+
 ?>
 <!-- Begin Page Content -->
 <main>
     <div class="container-fluid">
 
         <!-- Page Heading -->
-        <h1 class="h3 mb-2 text-gray-800"><?= $titulo ?></h1>
         <ol class="breadcrumb mb-4">
             <li class="breadcrumb-item"><a href="<?= base_url()?>">Inicio</a></li>
             <li class="breadcrumb-item active"><?= $titulo ?></li>
@@ -26,7 +27,10 @@
                         <thead>
                             <tr>
                                 <th>ID</th>
+                                <th>Código</th>
                                 <th>Nombre</th>
+                                <th>Precio</th>
+                                <th>Existencias</th>
                                 <th></th>
                                 <th></th>
 
@@ -34,8 +38,11 @@
                         </thead>
                         <tfoot>
                             <tr>
-                                <th>ID</th>
+                               <th>ID</th>
+                                <th>Código</th>
                                 <th>Nombre</th>
+                                <th>Precio</th>
+                                <th>Existencias</th>
                                 <th></th>
                                 <th></th>
                             </tr>
@@ -47,9 +54,12 @@
                             ?>
                                 <tr>
                                     <td><?php echo $dato['id']; ?></td>
-                                    <td><?php echo $dato['nombre']; ?></td>
-                                    <td><a class="btn btn-warning" href="<?= base_url() ?>productos/editar/<?php echo $dato['id']; ?>"><i class="fas fa-edit"></i></a></td>
-                                    <td><a class="btn btn-danger" href="<?= base_url() ?>productos/eliminar/<?php echo $dato['id']; ?>"><i class="fas fa-trash-alt"></a></td>
+                                    <td><?php echo $dato['codigo']; ?></td>
+                                     <td><?php echo $dato['nombre']; ?></td>
+                                      <td><?php echo number_to_currency($dato['precio_venta'], 'USD', 'en_US', 0); ?></td>
+                                       <td><?php echo $dato['existencias']; ?></td>
+                                    <td><a class="btn btn-warning btn-sm" href="<?= base_url() ?>productos/editar/<?php echo $dato['id']; ?>"><i class="fas fa-edit"></i></a></td>
+                                    <td><a data-toggle="modal" data-target="#modal-confirma" class="btn btn-danger btn-sm" href="#" data-href="<?= base_url() ?>productos/eliminar/<?php echo $dato['id']; ?>"><i class="fas fa-trash-alt"></i></a></td>
 
                                 </tr>
                             <?php
@@ -62,3 +72,24 @@
         </div>
     </div>
 </main>
+
+<!-- Modal confirmación -->
+    <div class="modal fade" id="modal-confirma" tabindex="-1">
+        <div class="modal-dialog modal-sm">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title">Eliminar Registro</h5>
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body">
+                    <p>¿Está seguro que desea eliminar el registro?</p>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Cancelar</button>
+                    <a type="button" class="btn btn-danger btn-ok">Eliminar</a>
+                </div>
+            </div>
+        </div>
+    </div>
